devex - removing webpack files for frontned
