--- web-client/src/index-public.dev.js
@@ -1,9 +1,0 @@
-import { appPublic } from './appPublic';
-import { applicationContextPublic } from './applicationContextPublic';
-
-/**
- * Initializes the app with dev environment context
- */
-const options = {};
-
-appPublic.initialize(applicationContextPublic, options);
