--- web-client/src/index-public.prod.js
@@ -1,7 +1,0 @@
-import { appPublic } from './appPublic';
-import { applicationContextPublic } from './applicationContextPublic';
-
-/**
- * Initializes the app with prod environment context
- */
-appPublic.initialize(applicationContextPublic);
